# server/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PTCF Room & Equipment Reservation System API",
    "description": "API documentation for the UPLB ICropS Plant Tissue Culture Facility reservation management system.",
    "version": "1.0.0",
    "contact": {
      "name": "PTCF Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Local development server"
    },
    {
      "url": "https://your-render-url.onrender.com/api",
      "description": "Production server (Render)"
    }
  ],
...
  "paths": {
...
    "/rooms": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "summary": "Get all rooms",
        "description": "Returns a list of all rooms. Accessible to all authenticated users.",
        "operationId": "getAllRooms",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of rooms retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "tags": [
          "Rooms"
        ],
        "summary": "Create new room",
        "description": "Creates a new room with optional image upload. Requires staff or admin role.",
        "operationId": "createRoom",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "location",
                  "capacity"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Room name",
                    "example": "Culture Room A"
                  },
                  "description": {
                    "type": "string",
                    "description": "Room description",
                    "example": "Primary tissue culture laboratory"
                  },
                  "location": {
                    "type": "string",
                    "description": "Physical location",
                    "example": "ICropS 2nd Floor"
                  },
                  "capacity": {
                    "type": "integer",
                    "description": "Maximum occupancy",
                    "example": 8
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "available",
                      "unavailable",
                      "maintenance"
                    ],
                    "description": "Availability status (optional, defaults to 'available')",
                    "example": "available"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file (optional, max 5MB, JPEG/PNG/GIF/WebP)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Room created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Missing required fields: name, description, location, and capacity are required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{id}": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "summary": "Get room by ID",
        "description": "Returns a single room by ID. Accessible to all authenticated users.",
        "operationId": "getRoomById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Room ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Room not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Room not found"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Rooms"
        ],
        "summary": "Update room",
        "description": "Updates an existing room. Can optionally update the image. Requires staff or admin role.",
        "operationId": "updateRoom",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Room ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Room name (optional)"
                  },
                  "description": {
                    "type": "string",
                    "description": "Room description (optional)"
                  },
                  "location": {
                    "type": "string",
                    "description": "Physical location (optional)"
                  },
                  "capacity": {
                    "type": "integer",
                    "description": "Maximum occupancy (optional)"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "available",
                      "unavailable",
                      "maintenance"
                    ],
                    "description": "Availability status (optional)"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "New image file (optional, replaces existing image)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Room updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "description": "Room not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Rooms"
        ],
        "summary": "Delete room",
        "description": "Deletes a room by ID. Requires staff or admin role.",
        "operationId": "deleteRoom",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Room ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Room deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "description": "Room not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid credentials"
          }
        }
      },
...
      "Room": {
        "type": "object",
        "description": "Room model schema",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "Room name"
          },
          "description": {
            "type": "string",
            "description": "Room description"
          },
          "location": {
            "type": "string",
            "description": "Physical location"
          },
          "capacity": {
            "type": "integer",
            "description": "Maximum occupancy"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "description": "Cloudinary image URL"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "unavailable",
              "maintenance"
            ],
            "description": "Availability status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized - Missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "missingToken": {
                "value": {
                  "message": "Missing Bearer token"
                }
              },
              "invalidToken": {
                "value": {
                  "message": "Invalid or expired token"
                }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Forbidden - Insufficient role (requires staff or admin)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Forbidden: insufficient role"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter ONLY your JWT token (without 'Bearer' prefix). Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      }
    }
  },
  "security": []
}